split up app.jsx
